fix(goals): fix the editing status for all sections
- Fixed a bug with simultaneous editing of all sections
- The logic of displaying goals has been decomposed (with/without sections)
