+++ components/goals/GoalItem.tsx
@@ -12,9 +12,10 @@ import { useGoalsStore } from '@/stores/goals-store';
 import useGlobalYear from '@/hooks/use-global-year';
 import useGoalYearSettings from '@/hooks/use-goal-year-settings';
 
-import StatusItem from '../StatusItem';
 import GoalsUpdateFormWrapper from './GoalsUpdateFormWrapper';
-import GoalsItemFooter from './goals-item-footer';
+import GoalsItemFooter from './GoalsItemFooter';
+
+import StatusItem from '../StatusItem';
 
 import type { Description } from '@/types/form-goal.types';
 import type { StatusKeys } from '@/types/status.types';
